requirements.txt
streamlit==1.28.1
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
Pillow==10.0.1